azCommunicationService/requirements.txt
Flask
azure-communication-email
tomli